ETL Folder restructuring, utils folder and architecture added, Seed ai tools loaded separately
