+++ .gitignore
@@ -33,4 +33,3 @@ build/
 # Operating system specific files (e.g., macOS, Windows)
 .DS_Store
 Thumbs.db
-etl/ai_tools_scraped.json
