+++ backend/app/core/database.py
@@ -25,7 +25,7 @@ DB_URL = f"postgresql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
 engine = create_engine(DB_URL)
 
 
-SessionLocal = sessionmaker(autoflush=False, atutocomit=False, bind=engine)
+SessionLocal = sessionmaker(autoflush=False, autocomit=False, bind=engine)
 
 Base = declarative_base()
