+++ etl/tests/test_models.py
@@ -1,0 +1,2 @@
+import pytest
+from idempotent_etl_job import read_data
